Don't stop dragging on mouse leave
It feels more robust if the dragging continues even if the mouse
leaves the area that contains the drag listener. This makes it easier to
pick colors near the edges (particularly near *two* edges, i.e. in the
corners)

For this to work, another listener had to be altered: the `mouseup`
listener is now registered on document.body, instead of the drag area.
It is also unregistered once the dragging is complete.